10 - Added global position variables for each servo
The global variables for servo positions was needed because servo.read() does NOT return the current position, but instead the last value written to the servo.  This causes initialization problems, and further problems when a write() is performed but the servo does not have time to move to that position.